refactor: hero-detail を features の下に配置
