+++ src/app/features/hero-detail/containers/hero-detail/hero-detail.component.css
